organized and optimized code and readablity
