+++ code/main.py
@@ -2,37 +2,15 @@ import datetime
 import speech_recognition as sr
 import pyttsx3
 from mood import (happy_list, sad_list, angry_list, depressed_list, suicidal_list, happy_mood_responses, sad_mood_responses, angry_mood_responses, depressed_mood_responses, suicidal_mood_responses, general_responses)
+from functions import (wishme, takecommand, speak_and_print, speak)
 
 # Initialize the text-to-speech engine
 engine = pyttsx3.init()
 voices = engine.getProperty("voices")
 engine.setProperty("voice", voices[1].id)
 
-# Function to speak the given audio
-def speak(audio):
-    engine.say(audio)
-    engine.runAndWait()
-
-# Function to greet the user based on the time of day
-def wishme():
-    hour = int(datetime.datetime.now().hour)
-    if 0 <= hour < 12:
-        print("Good morning!")
-        speak("Good morning!")
-    elif 12 <= hour < 18:
-        print("Good afternoon")
-        speak("Good afternoon!")
-    else:
-        print("Good evening!")
-        speak("Good evening!")
-
 # Call the wishme function to greet the user
 wishme()
-
-# Function to print a message and speak it
-def speak_and_print(message):
-    print(message)
-    speak(message)
 
 # Ask for the user's name and start the therapy session and desclaimer that say this is Not a real therapist just a script 
 while True:
@@ -40,14 +18,14 @@ while True:
 
     speak_and_print("The Thea development team emphasizes that while Thea aims to be supportive, it is not a licensed therapist. The AI is continually learning and evolving, and your feedback is valuable for its improvement. The Thea team does not take responsibility for any harm caused by the program; use it at your own risk.")
 
-    disclaimer = "If you understand that Thea is not a real therapist and you have read and understood the disclaimer, please type 'thea is not a real therapist and i have read and understood the disclaimer': "
+    disclaimer = "If you understand that Thea is not a real therapist and you have read and understood the disclaimer, please type 'y' to confirm: "
 
     
     speak_and_print(disclaimer)
     user_agree = input("type hear: ").lower()
     
-    if user_agree == "thea is not a real therapist and i have read and understood the disclaimer":
-        break  # Break out of the loop if the user agrees
+    if user_agree == "y":
+        break  # Break out of the loop if the user agrees to the terms of service
     elif user_agree == "admin code 110308":
         break
 
@@ -55,27 +33,8 @@ while True:
 speak_and_print("Now, please state your name for easy communication")
 name = input("Your Name: ")
 
-
-print(f"I am Thea, your therapist for today, {name} remember that you can end the session whenever you what by saying `ok, thank you for the session` to . Let's start. How are you feeling today?")
-speak(f"I am Thea, your therapist for today, {name} remember that you can end the session whenever you what by saying `ok, thank you for the session` to . Let's start. How are you feeling today")
-
-# Function to take microphone input and return string output
-def takecommand():    
-    r = sr.Recognizer()
-    with sr.Microphone() as source:
-        print("Listening...")
-        r.pause_threshold = 1
-        audio = r.listen(source)
-
-    try:
-        print("Recognizing...")
-        user_statement = r.recognize_google(audio, language="en-us")
-        print(f"User side: {user_statement}\n")
-
-    except Exception as e:
-        print("Say that again, please...")
-        return "None"
-    return user_statement
+#the session has started after the introduction
+speak_and_print(f"I am Thea, your therapist for today, {name} remember that you can end the session whenever you what by saying `ok thank you for the session` to . Let's start. How are you feeling today?")
 
 # Main program execution
 if __name__ == "__main__":
@@ -98,8 +57,9 @@ if __name__ == "__main__":
         elif any(word in user_statement for word in suicidal_list):
             suicidal_mood_responses()
             
-        elif user_statement == "ok thank you for the session":
+        elif user_statement == "ok thanks for the session":
             speak_and_print("Ok then see you next time")
+            break
             
         if all(word not in user_statement for word in happy_list + sad_list + angry_list + depressed_list + suicidal_list):
             general_responses()
